requirements.txt

langchain
langchain-google-genai
flask-cors==3.0.10
transformers==4.30.2
tensorflow==2.16.1
tf-keras  
langchain_community
python-dotenv==1.0.0
google-api-python-client==2.92.0